gcp-pulumi/requirements.txt
pulumi>=3.0.0,<4.0.0
pulumi-gcp>=7.0.0,<8.0.0
pulumi-github==6.2.0
grpcio==1.60.1
grpcio-status==1.60.1
python-dotenv==1.0.1
